style(web-components): refactor to TS
Split config functions into separate modules

Add type and language to docs.source
